chore: optimize code
