+++ apps/router-demo/router-remote1-2001/rsbuild.config.ts
@@ -38,12 +38,12 @@ export default defineConfig({
             './export-app': './src/export-App.tsx',
           },
           shared: {
-            react: {
-              singleton: true,
-            },
-            'react-dom': {
-              singleton: true,
-            },
+            // react: {
+            //   singleton: true,
+            // },
+            // 'react-dom': {
+            //   singleton: true,
+            // },
             // 'react-router-dom': {
             //   singleton: true,
             // },
